refactor: fsd refactoring organization across modules
Refactored file paths to improve module organization and consistency, renaming and relocating several files under clearer directories. Updated imports and exports to reflect the new structure, including adjustments in configuration for FSD rules and scripts in package.json. These changes aim to streamline development and maintain alignment with project architecture guidelines.
