+++ frontend/src/features/template/search/index.ts
@@ -1,5 +1,5 @@
-export * from './order-by-filter';
-export * from './order-direction-filter';
-export * from './query-filter';
-export * from './tags-filter';
-export * from './topics-filter';
+export * from './order-by';
+export * from './order-direction';
+export * from './query';
+export * from './tags';
+export * from './topics';
