+++ frontend/src/features/template/search/query/index.ts
