+++ frontend/src/features/template/search/topics/ui/index.tsx
