+++ frontend/src/shared/api/helpers/error-handler.ts
