+++ frontend/src/shared/api/index.ts
@@ -1,5 +1,5 @@
 export * from './types';
 export * from './schemes';
-export * from './lib';
+export * from './helpers';
 export * from './constants';
 export * from './types';
